refactor past event component to use gatsby static query hook
