+++ package.json
@@ -8,7 +8,7 @@
     "@types/react-google-recaptcha": "^1.0.0",
     "babel-plugin-styled-components": "^1.10.0",
     "date-fns": "^1.30.1",
-    "gatsby": "^2.1.9",
+    "gatsby": "^2.2.3",
     "gatsby-image": "^2.0.29",
     "gatsby-plugin-favicon": "^3.1.5",
     "gatsby-plugin-manifest": "^2.0.17",
